Add useBrowserOnChromeOS flag.
This will be set to true by default, and will indicate that the
developer hasn't added any new activities or services beyond what
come bundled with svgomg-twa. This ensures that there is no interaction
with Android that we don't know of, and can be used to interpret that
installing this web app as a Desktop PWA on ChromeOS won't break
user experience.